Abstracts sticky header into seperate component
